requirements.txt
apify-client==1.6.4
beautifulsoup4==4.12.3
gradio==4.21.0
huggingface-hub==0.20.1
langchain==0.1.16
openai==1.20.0
opencv-python==4.8.0.74
pandas==2.0.3
PEFT==0.7.1
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==1.3.1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2+cu117
torchvision==0.15.2
transformers==4.31.0
typing_extensions==4.7.1
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.3.6
xformers==0.0.22